% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCopAUC.R
\name{BiCopAUC2Par}
\alias{BiCopAUC2Par}
\title{Parameter of a bivariate copula for a given Kendall's tau value}
\usage{
BiCopAUC2Par(family, AUC, c0)
}
\arguments{
\item{family}{copula family. check \code{BiCopPar2Tau} for detail.}

\item{AUC}{the AUC value}

\item{c0}{prevalence; cutoff point of latent uniform parameter}
}
\description{
This function computes the parameter of a one parameter bivariate copula for a given value of AUC.
}
\examples{
# BiCopAUC2Par(3, 0.7274113, c0 = 0.5, lower = 1e-5, upper = 5)
}
