% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCopCind.R
\name{BiCopCind2Par}
\alias{BiCopCind2Par}
\title{Parameter of a bivariate copula for a given C-index value}
\usage{
BiCopCind2Par(family, Cind)
}
\arguments{
\item{family}{copula family. check \code{BiCopPar2Tau} for detail.}

\item{Cind}{value (numeric in [0.5, 1])}
}
\description{
This function computes the parameter of a one parameter bivariate copula for a given value of C-index.
}
\examples{
# BiCopCstat2Par(3, 0.8)
}
