% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCopHfunc.R
\name{BiCopHfunc}
\alias{BiCopHfunc}
\title{Function from CDVine
https://github.com/cran/CDVine}
\usage{
BiCopHfunc(u1, u2, family, par, par2 = 0)
}
\description{
Function from CDVine
https://github.com/cran/CDVine
}
