% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicopiAUC.R
\name{BiCopiAUC2Par}
\alias{BiCopiAUC2Par}
\title{Parameter of a bivariate copula for a given Kendall's tau value}
\usage{
BiCopiAUC2Par(family, iAUC)
}
\arguments{
\item{family}{copula family. check \code{BiCopPar2Tau} for detail.}

\item{iAUC}{the integrated AUC value}

\item{c0}{prevalence; cutoff point of latent uniform parameter}
}
\description{
This function computes the parameter of a one parameter bivariate copula for a given value of AUC.
}
\examples{
# BiCopiAUC2Par(3, 0.75)
}
