% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCopAUC.R
\name{BiCopPar2AUC}
\alias{BiCopPar2AUC}
\title{AUC value of a bivariate copula}
\usage{
BiCopPar2AUC(family, par, par2 = 0, c0, type = "cumulative")
}
\arguments{
\item{family}{copula family. check \code{BiCopPar2Tau} for detail.}

\item{par}{Copula parameter.}

\item{par2}{Second parameter for the two parameter BB1, BB6, BB7 and BB8 copulas (default: \code{par2 = 0}).}

\item{c0}{prevalence; cutoff point of latent uniform parameter}

\item{type}{type is cumulative or incidence}
}
\description{
This function computes the theoretical AUC value of a bivariate copula for given parameter values.
}
\examples{
# BiCopPar2AUC(3, 1, c0 = 0.5)
}
