% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCopCind.R
\name{BiCopPar2Cind}
\alias{BiCopPar2Cind}
\title{C-index value of a bivariate copula}
\usage{
BiCopPar2Cind(family, par, par2 = 0)
}
\arguments{
\item{family}{copula family. check \code{BiCopPar2Tau} for detail.}

\item{par}{Copula parameter.}

\item{par2}{Second parameter for the two parameter BB1, BB6, BB7 and BB8 copulas (default: \code{par2 = 0}).}

\item{c0}{prevalence; cutoff point of latent uniform parameter}
}
\description{
This function computes the theoretical C-index value of a bivariate copula for given parameter values.
}
\examples{
BiCopPar2AUC(3, 1, c0 = 0.5)
}
