% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicopiAUC.R
\name{BiCopPar2iAUC}
\alias{BiCopPar2iAUC}
\title{iAUC value of a bivariate copula}
\usage{
BiCopPar2iAUC(
  family,
  par,
  par2 = 0,
  show_AUC = T,
  c0_length = 50,
  weight = NULL,
  q_density = NULL,
  d_density = NULL,
  type = "cumulative"
)
}
\arguments{
\item{family}{copula family. check \code{BiCopPar2Tau} for detail.}

\item{par}{Copula parameter.}

\item{par2}{Second parameter for the two parameter BB1, BB6, BB7 and BB8 copulas (default: \code{par2 = 0}).}

\item{show_AUC}{display time depednent AUC (default: TRUE)}

\item{c0_length}{length of cutoff points}

\item{weight}{weight of time dependent AUC}

\item{q_density}{marginal quantile function of Time distribution}

\item{d_density}{marginal PDF of Time distribution}

\item{type}{type is cumulative or incidence}
}
\description{
This function computes the theoretical AUC value of a bivariate copula for given parameter values.
}
\examples{
#BiCopPar2iAUC( family = 3, par = 1)
}
