% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCopPar2Tau.R
\name{BiCopPar2Tau}
\alias{BiCopPar2Tau}
\title{Function from CDVine
https://github.com/cran/CDVine}
\usage{
BiCopPar2Tau(family, par, par2 = 0)
}
\description{
Function from CDVine
https://github.com/cran/CDVine
}
