% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDVineSim.R
\name{CDVineSim}
\alias{CDVineSim}
\title{Function from CDVine
https://github.com/cran/CDVine}
\usage{
CDVineSim(N, family, par, par2 = rep(0, length(family)), type)
}
\description{
Function from CDVine
https://github.com/cran/CDVine
}
