% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCPP.R
\name{simuCCP_para}
\alias{simuCCP_para}
\title{Simulate Concordance Probability based on bivariate copula with known family and parameters}
\usage{
simuCCP_para(N, metric, family, par, c0 = 0.5)
}
\arguments{
\item{N}{sample size}

\item{metric}{simulate metric (C-index, iAUC, AUC)}

\item{family}{A d*(d-1)/2 integer vector of C-vine pair-copula families (detail is in \code{CDVineSim})}

\item{par}{A d*(d-1)/2 vector of pair-copula parameters.}

\item{c0}{prevalence; cutoff point of latent uniform parameter (Used in AUC)}
}
\description{
This function generate multivariate uniform distribution with designed copula family and
parameters (C-index, iAUC, AUC)
}
